+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [develop]
+
+### Added
+- UI element `ListBox` to display multiple selectable items
+- Icon for subtitles
+- Icon for audio tracks
+- Demo section within the UI variants in the playground
+- Demo with separate `SettingsPanel`s for subtitles and audio tracks
+
 ## [2.17.1]
 
 ### Fixed